refactor(s1): move faucet orchestration to dkg-core/faucet-orchestration
S1 step 3 of execution-plan.md §3.S1 (issue #386 — Hermes setup parity).
Extract `readWallets`, `readWalletsWithRetry`, `logManualFundingInstructions`,
`sleep`, and a new orchestrator `fundWalletsBestEffort({ network, callerId,
didStartDaemon })` from adapter-openclaw to
@origintrail-official/dkg-core/faucet-orchestration.ts so adapter-hermes
can reuse them in S2 (issue #386 acceptance criterion: "--no-fund truly
means do not perform faucet funding"; H-AC-19 will assert the 5×1s retry
semantics in S2).

- Added: packages/core/src/faucet-orchestration.ts with `readWallets`,
  `readWalletsWithRetry`, `logManualFundingInstructions`,
  `fundWalletsBestEffort` (orchestrator wrapping the faucet URL check,
  retry-after-daemon-start, requestFaucetFunding call, and manual-curl
  fallback that the OpenClaw runSetup body used inline)
- Added: exports via packages/core/src/index.ts
- adapter-openclaw/src/setup.ts now thin-re-exports the four helpers and
  refactors its runSetup Step 6 block to call `fundWalletsBestEffort`
  instead of inlining the orchestration. The body of the orchestrator is
  lifted line-for-line from the previous inline block — behavior-equivalent
- adapter-openclaw/test/setup.test.ts: dual-mock pattern for
  `requestFaucetFunding` (barrel + dkg-core/dist/faucet.js), since
  `fundWalletsBestEffort` inside core calls `requestFaucetFunding` via
  intra-package import that the barrel mock alone wouldn't intercept.
  Same vitest mock-boundary pattern used in S1.1 + S1.2

The 5×1s retry semantics in `readWalletsWithRetry` are preserved exactly.
Faucet failures stay non-fatal; the orchestrator never throws.

Test results vs s1-baseline.md (post-extraction):
- @origintrail-official/dkg-core: 33 files, 550 tests — green
- @origintrail-official/dkg-adapter-openclaw: 22 files, 974 tests —
  same 2 pre-existing failures as baseline; zero new failures
- @origintrail-official/dkg-adapter-hermes: 1 file, 60 tests — green
- @origintrail-official/dkg-node-ui: 20 files — green
- @origintrail-official/dkg (CLI): curated subset gate per s1-baseline.md
